requirements.txt
flask
pillow
torch
transformers
numpy